sdpa -> pa renaming
