+++ src/common/transformations/tests/op_conversions/sdpa_to_paged_attention_test.cpp
@@ -59,11 +59,11 @@
 #include "openvino/op/util/variable.hpp"
 #include "openvino/op/variadic_split.hpp"
 #include "openvino/pass/visualize_tree.hpp"
+#include "transformations/paged_attention/position_ids_replacer.hpp"
+#include "transformations/paged_attention/prev_sequence_length_pattern.hpp"
+#include "transformations/paged_attention/state_management_pattern.hpp"
+#include "transformations/paged_attention/total_sequence_length_pattern.hpp"
 #include "transformations/rt_info/keep_const_precision.hpp"
-#include "transformations/sdpa_to_paged_attention/position_ids_replacer.hpp"
-#include "transformations/sdpa_to_paged_attention/prev_sequence_length_pattern.hpp"
-#include "transformations/sdpa_to_paged_attention/state_management_pattern.hpp"
-#include "transformations/sdpa_to_paged_attention/total_sequence_length_pattern.hpp"
 #include "transformations/utils/gen_pattern.hpp"
 
 using namespace ov;
